requirements.txt
datetime
paho-mqtt
python-dotenv
pyinstaller